I have begun integrating NGR into the PSR notebook. This is going to be quite a challenge.
I have to make a lot of changes to the code, but I've set up a separate notebook to store the findings. As of right now, the Monte Carlo simluations are all quite skewed (especially the Linearisation one). i am not sure if this is an error or if it's working, and that NGR behaves weirdly for large volume of trades.

Tomorrow do some rigorous testing (compare side by side with the NGRanalysis notebok to check for errors)